migrate_schema.py
import os
from dotenv import load_dotenv
import httpx

load_dotenv("pantheon.env", override=True)
url = os.environ["SUPABASE_URL"]
key = os.environ["SUPABASE_SERVICE_ROLE_KEY"]

def alter_table():
    endpoint = f"{url}/rest/v1/rpc/exec_sql"
    sql = """
    ALTER TABLE agent_genomes ADD COLUMN IF NOT EXISTS religion TEXT;
    ALTER TABLE agent_genomes ADD COLUMN IF NOT EXISTS cultural_background TEXT;
    """
    
    # Supabase REST API doesn't expose a raw exec_sql by default unless we created the function.
    # We can try to just use standard raw SQL execution. Wait, is there a way to run DDL from python client?
    pass
