requirements.txt
modal
supabase
anthropic
pydantic
python-dotenv
apify-client
requests
python-docx
streamlit
PyPDF2
